+++ tests/test_examples_app.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 #
 # This file is part of Invenio.
-# Copyright (C) 2016 CERN.
+# Copyright (C) 2016, 2017 CERN.
 #
 # Invenio is free software; you can redistribute it
 # and/or modify it under the terms of the GNU General Public License as
@@ -28,6 +28,7 @@ import os
 import signal
 import subprocess
 import time
+from os.path import abspath, dirname, join
 
 import pytest
 
@@ -36,31 +37,32 @@ import pytest
 def example_app():
     """Example app fixture."""
     current_dir = os.getcwd()
-    # go to example directory
-    project_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-    exampleappdir = os.path.join(project_dir, 'examples')
-    os.chdir(exampleappdir)
-    # setup example
-    cmd = './app-setup.sh'
-    exit_status = subprocess.call(cmd, shell=True)
-    assert exit_status == 0
-    # setup example
-    cmd = './app-fixtures.sh'
-    exit_status = subprocess.call(cmd, shell=True)
-    assert exit_status == 0
-    # Starting example web app
-    cmd = 'FLASK_APP=app.py flask run --debugger -p 5000'
-    webapp = subprocess.Popen(cmd, stdout=subprocess.PIPE,
-                              preexec_fn=os.setsid, shell=True)
-    time.sleep(10)
-    # return webapp
+
+    # Go to example directory
+    project_dir = dirname(dirname(abspath(__file__)))
+    exampleapp_dir = join(project_dir, 'examples')
+    os.chdir(exampleapp_dir)
+
+    # Setup application
+    assert subprocess.call('./app-setup.sh', shell=True) == 0
+
+    # Setup fixtures
+    assert subprocess.call('./app-fixtures.sh', shell=True) == 0
+
+    # Start example app
+    webapp = subprocess.Popen(
+        'FLASK_APP=app.py flask run --debugger -p 5000',
+        stdout=subprocess.PIPE, preexec_fn=os.setsid, shell=True)
+    time.sleep(3)
     yield webapp
-    # stop server
+
+    # Stop server
     os.killpg(webapp.pid, signal.SIGTERM)
-    # tear down example app
-    cmd = './app-teardown.sh'
-    subprocess.call(cmd, shell=True)
-    # return to the original directory
+
+    # Tear down example app
+    subprocess.call('./app-teardown.sh', shell=True)
+
+    # Return to the original directory
     os.chdir(current_dir)